refactor(Purchaser): Purchaser를 만들면서 LottoVendingMachine이 로또 개수까지 계산할 필요가 없게 됨 단순히 로또만 생성하면 된다.
